backend/requirements.txt
flask
flask-cors
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
langchain
langchain-google-genai
nltk
google-generativeai